kairos/backend/requirements.txt
# FastAPI 및 웹 서버 관련
fastapi==0.103.1
uvicorn==0.23.2
starlette==0.27.0
anyio==3.7.1
h11==0.14.0
sniffio==1.3.1
python-multipart==0.0.6
requests==2.31.0

# 데이터베이스 및 모델링
pydantic==2.5.2
pydantic_core==2.14.5
pydantic-settings==2.1.0
annotated-types==0.7.0
pandas==2.1.3
numpy==1.26.4

# 보안 및 인증
python-jose==3.3.0
PyJWT==2.8.0
passlib==1.7.4
bcrypt==4.0.1
cryptography==41.0.3
ecdsa==0.19.1
pyasn1==0.6.1
rsa==4.9

# 환경 설정
python-dotenv==1.0.0
PyYAML==6.0.1

# 유틸리티
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
cffi==1.17.1
exceptiongroup==1.2.2
idna==3.10
typing_extensions==4.12.2
urllib3==2.3.0
six==1.17.0
pycparser==2.22

# 날짜 및 시간
python-dateutil==2.9.0.post0
pytz==2025.1
tzdata==2025.1

# 테크니컬 인디케이터
ta-lib==0.6.3
